requirements.txt
# Core dependencies
instagram-private-api==1.6.0  # Instagram API handling
python-dotenv==0.19.2        # Environment variables
keyboard==0.13.5             # Keyboard monitoring

# Database
sqlite3                      # Built into Python

# Testing
pytest==7.3.1               # Testing framework
pytest-cov==4.1.0           # Test coverage reporting
pytest-mock==3.10.0         # Mocking for tests

# Utilities
typing-extensions==4.5.0    # Type hinting support
python-dateutil==2.8.2      # Date handling
requests==2.31.0            # HTTP requests
urllib3==1.26.15            # HTTP client
certifi>=2023.7.22          # Security certificates

# Optional but recommended
black==23.3.0               # Code formatting
flake8==6.0.0              # Code linting
mypy==1.3.0                # Type checking